requirements.txt
dash
dash-bootstrap-components
pandas
plotly
requests
fastapi
uvicorn[standard]
snowflake-connector-python
pymongo
prophet
scikit-learn
python-dotenv